# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Exactamente API",
    "description": "Recursos universitarios de Exactas: parciales, resúmenes y finales.\n\nEste documento se genera desde los schemas Zod del backend, así que no\npuede desincronizarse de la validación. Los clientes (web, panel admin y\nservidor MCP) generan sus tipos desde acá.\n\n**Paginación:** `page` (default 1) y `limit` (default 20, máx 100).\nExcepción: `GET /careers` devuelve `{ data: [...] }` sin paginar.\n\n**Auth:** las rutas protegidas piden `Authorization: Bearer <jwt>`.\nLas de `/admin/*` requieren rol `admin` o `superadmin`.",
    "version": "1.0.0",
    "license": {
      "name": "MIT",
      "url": "https://github.com/exactamente-ar/exactamente-backend"
    }
  },
  "servers": [
    {
      "url": "https://api.exactamente.com.ar/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "bearer",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserRole": {
        "id": "UserRole",
        "type": "string",
        "enum": [
          "superadmin",
          "admin",
          "user"
        ]
      },
      "PublicUser": {
        "id": "PublicUser",
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
          },
          "displayName": {
            "type": "string"
          },
          "photoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "role": {
            "description": "Los registros nuevos siempre son user",
            "$ref": "#/components/schemas/UserRole"
          },
          "emailVerified": {
            "description": "Existe en DB pero hoy ningún flujo de producto lo usa",
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "email",
          "displayName",
          "photoUrl",
          "role",
          "emailVerified",
          "createdAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
...
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "postApiV1AuthRegister",
        "tags": [
          "Auth"
        ],
        "summary": "Registrar un usuario",
        "description": "Los registros nuevos siempre reciben rol `user`. Límite: 5 por hora por IP.",
        "responses": {
          "201": {
            "description": "Usuario creado con su token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/PublicUser"
                    },
                    "token": {
                      "description": "JWT — mandarlo como `Authorization: Bearer <token>`",
                      "type": "string"
                    }
                  },
                  "required": [
                    "user",
                    "token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Conflicto de estado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "postApiV1AuthLogin",
        "tags": [
          "Auth"
        ],
        "summary": "Iniciar sesión",
        "description": "Rechaza con 401 a los usuarios creados vía Google, que no tienen contraseña. Límite: 10 cada 15 min por IP.",
        "responses": {
          "200": {
            "description": "Sesión iniciada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/PublicUser"
                    },
                    "token": {
                      "description": "JWT — mandarlo como `Authorization: Bearer <token>`",
                      "type": "string"
                    }
                  },
                  "required": [
                    "user",
                    "token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "getApiV1AuthMe",
        "tags": [
          "Auth"
        ],
        "summary": "Usuario autenticado actual",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Usuario de la sesión",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "$ref": "#/components/schemas/PublicUser"
                    }
                  },
                  "required": [
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google": {
      "get": {
        "operationId": "getApiV1AuthGoogle",
        "tags": [
          "Auth"
        ],
        "summary": "Iniciar el flujo OAuth de Google",
        "responses": {
          "302": {
            "description": "Redirige al consent screen de Google"
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/google/callback": {
      "get": {
        "operationId": "getApiV1AuthGoogleCallback",
        "tags": [
          "Auth"
        ],
        "summary": "Callback de Google",
        "description": "Canjea el code por un JWT propio y redirige al frontend con un código de un solo uso. Ante error redirige a `/upload?error=...`.",
        "responses": {
          "302": {
            "description": "Redirige al frontend"
          },
          "429": {
            "description": "Rate limit excedido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/exchange": {
      "post": {
        "operationId": "postApiV1AuthExchange",
        "tags": [
          "Auth"
        ],
        "summary": "Canjear el código de un solo uso por un JWT",
        "description": "Cierra el flujo OAuth sin exponer el JWT en la URL.",
        "responses": {
          "200": {
            "description": "JWT",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validación fallida",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sin token o token inválido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "description": "Mensaje descriptivo, apto para mostrar al usuario",
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}